python-ai-services/requirements.txt
flask
pymongo[srv]
python-dotenv
scipy
requests